Addresses concern with redundant data structures in validation results.
Previously, validation results were packaged as a vector inside an option.
This was an unnecessary level of encapsulation as there are currently
only single errors returned at a time and a zero length vector could
have been sufficient to signify an absence of errors. This commit
simplifies the error packaging into an option of a single error
instance.