Discards comments and catches superfluous inline content.
Prior to this commit, comments were treated as literal nodes for matching purposes. Also, there was an
ommision that allowed superfluous inline content to slip through validation resulting in false
positives. This commit changes the validation logic to discard all html when matching and closes the
loophole for inline content.