Updates the README and README prescription licensing section
Previously, the licensing section did not validate because inline content prompts will not consume
heterogeneous inline content types. A -!!- token within a block of textual content will therefore
not consume an adjacent link. This is a known limitation of the current implementation. This commit
alters the licensing section to allow for multiple copyright holders to be specified in mailto
link format.