Fixes bug with validating optional inline content
Previously, there was a bug that caused false negatives to occur when validating documents with inline elements composed
of heterogeneous types where some of the elements were strictly optional. The fix for this was to adopt the same backtracking
algorithm used for validating block elements for inline elements. This commit applies that change.