Adds a fail-early option
Previously, the --pharmacy option would validate every document listed in the pharmacy file before
reporting results. In some circumstances, the desired behavior may be to stop as soon as a problem
is encountered. This commit adds a --fail-early option to be used in conjuction with the --pharmacy
option to trigger that behavior.