Reverse the validation order for block elements
Prior to this commit, the algorithm for validating block elements worked
from the end of the document to the beginning. This was because it was
simpler to process repeatable tokens in that direction. This caused
unintuitive error reporting however, since for humans it makes more
sense to evaluate the document from beginning to end. This commit refactors
the block validation logic to be performed starting at the beginning of
the document which improves the error reporting.